# Conflicts:
#	packages/@angular/cli/blueprints/ng/files/__path__/app/app.component.html
#	packages/@angular/cli/blueprints/ng/files/__path__/app/app.component.spec.ts
#	packages/@angular/cli/blueprints/ng/files/__path__/app/app.component.ts
#	packages/@angular/cli/blueprints/ng/files/__path__/app/app.module.ts
#	packages/@angular/cli/blueprints/ng/files/package.json
#	packages/@angular/cli/blueprints/ng/index.ts
#	packages/@angular/cli/tasks/init.ts

+++ packages/@angular/cli/tasks/serve.ts
@@ -31,7 +31,9 @@ export default Task.extend({
     if (projectConfig.project && projectConfig.project.ejected) {
       throw new SilentError('An ejected project cannot use the build command anymore.');
     }
-    rimraf.sync(path.resolve(this.project.root, outputPath));
+    if (serveTaskOptions.deleteOutputPath) {
+      rimraf.sync(path.resolve(this.project.root, outputPath));
+    }
 
     const serveDefaults = {
       // default deployUrl to '' on serve to prevent the default from .angular-cli.json
@@ -44,12 +46,21 @@ export default Task.extend({
 
     const serverAddress = url.format({
       protocol: serveTaskOptions.ssl ? 'https' : 'http',
-      hostname: serveTaskOptions.host,
+      hostname: serveTaskOptions.host === '0.0.0.0' ? 'localhost' : serveTaskOptions.host,
       port: serveTaskOptions.port.toString()
     });
+
+    if (serveTaskOptions.disableHostCheck) {
+      ui.writeLine(oneLine`
+          ${chalk.yellow('WARNING')} Running a server with --disable-host-check is a security risk.
+          See https://medium.com/webpack/webpack-dev-server-middleware-security-issues-1489d950874a
+          for more information.
+        `);
+    }
+
     let clientAddress = serverAddress;
-    if (serveTaskOptions.liveReloadClient) {
-      const clientUrl = url.parse(serveTaskOptions.liveReloadClient);
+    if (serveTaskOptions.publicHost) {
+      const clientUrl = url.parse(serveTaskOptions.publicHost);
       // very basic sanity check
       if (!clientUrl.host) {
         return Promise.reject(new SilentError(`'live-reload-client' must be a full URL.`));
@@ -75,7 +86,6 @@ export default Task.extend({
         ui.writeLine('  for information on working with HMR for Webpack.');
         entryPoints.push('webpack/hot/dev-server');
         webpackConfig.plugins.push(new webpack.HotModuleReplacementPlugin());
-        webpackConfig.plugins.push(new webpack.NamedModulesPlugin());
         if (serveTaskOptions.extractCss) {
           ui.writeLine(oneLine`
             ${chalk.yellow('NOTICE')} (HMR) does not allow for CSS hot reload when used
@@ -95,7 +105,7 @@ export default Task.extend({
       webpackConfig.plugins.unshift({
         apply: (compiler: any) => {
           compiler.plugin('after-environment', () => {
-            compiler.watchFileSystem = { watch: () => {} };
+            compiler.watchFileSystem = { watch: () => { } };
           });
         }
       });
@@ -148,8 +158,13 @@ export default Task.extend({
         poll: serveTaskOptions.poll
       },
       https: serveTaskOptions.ssl,
-      overlay: serveTaskOptions.target === 'development',
-      contentBase: false
+      overlay: {
+        errors: serveTaskOptions.target === 'development',
+        warnings: false
+      },
+      contentBase: false,
+      public: serveTaskOptions.publicHost,
+      disableHostCheck: serveTaskOptions.disableHostCheck
     };
 
     if (sslKey != null && sslCert != null) {
@@ -158,6 +173,11 @@ export default Task.extend({
     }
 
     webpackDevServerConfiguration.hot = serveTaskOptions.hmr;
+
+    // set publicPath property to be sent on webpack server config
+    if (serveTaskOptions.deployUrl) {
+      webpackDevServerConfiguration.publicPath = serveTaskOptions.deployUrl;
+    }
 
     if (serveTaskOptions.target === 'production') {
       ui.writeLine(chalk.red(stripIndents`
@@ -172,7 +192,8 @@ export default Task.extend({
 
     ui.writeLine(chalk.green(oneLine`
       **
-      NG Live Development Server is running on ${serverAddress}
+      NG Live Development Server is listening on ${serveTaskOptions.host}:${serveTaskOptions.port},
+      open your browser on ${serverAddress}
       **
     `));
 
@@ -183,7 +204,7 @@ export default Task.extend({
           return reject(err);
         }
         if (serveTaskOptions.open) {
-            opn(serverAddress);
+          opn(serverAddress);
         }
       });
     })
